increase timeout, check dupe first, refactor
